Now compiles and links airy_maker in VS 2005.
2007-07-29  Russell M. Taylor II  <taylorr@cs.unc.edu>

	* add_noise_to_image.cpp : Moving the "_LIB" definition into
		all of the locations that include magick/api.h to let
		us link statically without defining things in the project
		file.  Also fixing nits that let this compile without
		warnings on VS 2005.
	* airy_maker.vcproj : Same.
	* base_camera_server.cpp : Same.
	* file_stack_server.cpp : Same.
	* video.sln : Same.

+++ stocc_random_number_generator/testpois.cpp
@@ -41,7 +41,7 @@ int main() {
   int32 x1, x2;                        // category range
   
   // make random library
-  int32 seed = time(0);                // random seed
+  int32 seed = static_cast<int32>(time(0));                // random seed
   StochasticLib1 sto(seed);            // make instance of random library
 
   // calculate mean and variance
